feat(throttle): add higher-order lettable version of throttle
