+++ src/operator/throttle.ts
@@ -1,21 +1,5 @@
-import { Operator } from '../Operator';
 import { Observable, SubscribableOrPromise } from '../Observable';
-import { Subscriber } from '../Subscriber';
-import { Subscription, TeardownLogic } from '../Subscription';
-
-import { OuterSubscriber } from '../OuterSubscriber';
-import { InnerSubscriber } from '../InnerSubscriber';
-import { subscribeToResult } from '../util/subscribeToResult';
-
-export interface ThrottleConfig {
-  leading?: boolean;
-  trailing?: boolean;
-}
-
-export const defaultThrottleConfig: ThrottleConfig = {
-  leading: true,
-  trailing: false
-};
+import { throttle as higherOrder, ThrottleConfig, defaultThrottleConfig } from '../operators/throttle';
 
 /**
  * Emits a value from the source Observable, then ignores subsequent source
@@ -60,100 +44,5 @@ export const defaultThrottleConfig: ThrottleConfig = {
 export function throttle<T>(this: Observable<T>,
                             durationSelector: (value: T) => SubscribableOrPromise<number>,
                             config: ThrottleConfig = defaultThrottleConfig): Observable<T> {
-  return this.lift(new ThrottleOperator(durationSelector, config.leading, config.trailing));
-}
-
-class ThrottleOperator<T> implements Operator<T, T> {
-  constructor(private durationSelector: (value: T) => SubscribableOrPromise<number>,
-              private leading: boolean,
-              private trailing: boolean) {
-  }
-
-  call(subscriber: Subscriber<T>, source: any): TeardownLogic {
-    return source.subscribe(
-      new ThrottleSubscriber(subscriber, this.durationSelector, this.leading, this.trailing)
-    );
-  }
-}
-
-/**
- * We need this JSDoc comment for affecting ESDoc
- * @ignore
- * @extends {Ignored}
- */
-class ThrottleSubscriber<T, R> extends OuterSubscriber<T, R> {
-  private throttled: Subscription;
-  private _trailingValue: T;
-  private _hasTrailingValue = false;
-
-  constructor(protected destination: Subscriber<T>,
-              private durationSelector: (value: T) => SubscribableOrPromise<number>,
-              private _leading: boolean,
-              private _trailing: boolean) {
-    super(destination);
-  }
-
-  protected _next(value: T): void {
-    if (this.throttled) {
-      if (this._trailing) {
-        this._hasTrailingValue = true;
-        this._trailingValue = value;
-      }
-    } else {
-      const duration = this.tryDurationSelector(value);
-      if (duration) {
-        this.add(this.throttled = subscribeToResult(this, duration));
-      }
-      if (this._leading) {
-        this.destination.next(value);
-        if (this._trailing) {
-          this._hasTrailingValue = true;
-          this._trailingValue = value;
-        }
-      }
-    }
-  }
-
-  private tryDurationSelector(value: T): SubscribableOrPromise<any> {
-    try {
-      return this.durationSelector(value);
-    } catch (err) {
-      this.destination.error(err);
-      return null;
-    }
-  }
-
-  protected _unsubscribe() {
-    const { throttled, _trailingValue, _hasTrailingValue, _trailing } = this;
-
-    this._trailingValue = null;
-    this._hasTrailingValue = false;
-
-    if (throttled) {
-      this.remove(throttled);
-      this.throttled = null;
-      throttled.unsubscribe();
-    }
-  }
-
-  private _sendTrailing() {
-    const { destination, throttled, _trailing, _trailingValue, _hasTrailingValue } = this;
-    if (throttled && _trailing && _hasTrailingValue) {
-      destination.next(_trailingValue);
-      this._trailingValue = null;
-      this._hasTrailingValue = false;
-    }
-  }
-
-  notifyNext(outerValue: T, innerValue: R,
-             outerIndex: number, innerIndex: number,
-             innerSub: InnerSubscriber<T, R>): void {
-    this._sendTrailing();
-    this._unsubscribe();
-  }
-
-  notifyComplete(): void {
-    this._sendTrailing();
-    this._unsubscribe();
-  }
+  return higherOrder(durationSelector, config)(this);
 }
